+++ src/app/components/message/layout/Base.tsx
@@ -3,10 +3,13 @@ import classNames from 'classnames';
 import * as css from './layout.css';
 
 export const MessageBase = as<'div', css.MessageBaseVariants>(
-  ({ className, highlight, selected, collapse, autoCollapse, space, ...props }, ref) => (
+  (
+    { className, highlight, notifyHighlight, selected, collapse, autoCollapse, space, ...props },
+    ref
+  ) => (
     <div
       className={classNames(
-        css.MessageBase({ highlight, selected, collapse, autoCollapse, space }),
+        css.MessageBase({ highlight, notifyHighlight, selected, collapse, autoCollapse, space }),
         className
       )}
       {...props}