+++ src/app/pages/client/ClientLayout.tsx
@@ -1,5 +1,6 @@
 import { ReactNode } from 'react';
 import { Box } from 'folds';
+import { NotificationBanner } from '$components/notification-banner';
 
 type ClientLayoutProps = {
   nav: ReactNode;
@@ -8,6 +9,7 @@ type ClientLayoutProps = {
 export function ClientLayout({ nav, children }: ClientLayoutProps) {
   return (
     <Box grow="Yes">
+      <NotificationBanner />
       <Box shrink="No">{nav}</Box>
       <Box grow="Yes">{children}</Box>
     </Box>